Complete rewrite after a year of inactivity. I'll probably have to force-push this.
This commit sets up all the basic data structures to power the new API. It allows users to use the Shader object to parse arbitrary WGSL or GLSL code, compile it to a Naga Module, automatically bind the Uniforms and save them in a key-value store, and accept input by name. This version compiles with some warnings - basically unused symbols - but the demo is not ready yet. So while the basic structures are in place, this is not ready for tagging yet.