refactor Renderer, fix particles_splat example
this one is an important milestone. I feel confident to use it for more advanced use cases now. This could be a release candidate if I had not to throughly test JS with graceful WebGL fallback. Last time I ran the healthchecks there was a panic. Python worked well last time, but I have to test the new features on it.